requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient>=2.2.0